# Cache the validity mask, data pointers, and the chunk size