--- a/vector_setters.go
+++ b/vector_setters.go
@@ -19,9 +19,7 @@ const secondsPerDay = 24 * 60 * 60
 type fnSetVectorValue func(vec *vector, rowIdx C.idx_t, val any)
 
 func (vec *vector) setNull(rowIdx C.idx_t) {
-	C.duckdb_vector_ensure_validity_writable(vec.duckdbVector)
-	mask := C.duckdb_vector_get_validity(vec.duckdbVector)
-	C.duckdb_validity_set_row_invalid(mask, rowIdx)
+	C.duckdb_validity_set_row_invalid(vec.mask, rowIdx)
 
 	if vec.duckdbType == C.DUCKDB_TYPE_STRUCT {
 		for i := 0; i < len(vec.childVectors); i++ {
@@ -31,8 +29,7 @@ func (vec *vector) setNull(rowIdx C.idx_t) {
 }
 
 func setPrimitive[T any](vec *vector, rowIdx C.idx_t, v T) {
-	ptr := C.duckdb_vector_get_data(vec.duckdbVector)
-	xs := (*[1 << 31]T)(ptr)
+	xs := (*[1 << 31]T)(vec.ptr)
 	xs[rowIdx] = v
 }